Changed parameters for function calls which figure out the buffer size for H5Dread/H5Dwrite to use to allow the I/O calls to break up a user's buffer into pieces that are at least as large as the sequence of bytes being written in the fastest changing dimension.
Also fixed a hard-to-find bug in the hyperslab I/O routines which could cause
data to be corrupted when writing out fields to compound datatype data with
background preservation turned on and hyperslabs which were large enough to
require two I/O passes on a hyperslab block.  A pretty obscure situation, but
it would be worthwhile for users to upgrade to this code in order to be certain
that correct data is being written.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@1210 dab4d1a6-ed17-0410-a064-d1ae371a2980