started to bring the tools lib and the dumputil file together.  incorportated all the constants and from dumputil to to the tools lib and merged h5dump_fixtype together.  I had to add an extra parameter to the function to specify what action should be taken when it hits H5T_STRING in the case statement.  if the parameter (strDUAction) is true then it does what it use to do in dumputil but if it is false then it will do what it originally did in this file.
I also put the h5dump_sprint functions into the same file.  it isn't really merged yet.
i just wanted to get everything into a file and then worry about the merging of this
function.  The toolslib way of h5dump_sprint takes an extra parameter.  If this
parameter is set to NULL then it will do the dumputil's version of this function.
like i said this is just the first step i am taking. it will be merged better later.

the rest of the changes to the other files were related to the above changes.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@1253 dab4d1a6-ed17-0410-a064-d1ae371a2980