updated the functions that create the tstr.h5 and tdset.h5 so they have different values in the integer parts of the data.
for the tdset: it is i+j for the values instead of just i.(i and j are variables
used in a set of nested for loops)

for the tstr.h5: changed the data(originally just l to (l+k+j) squared).
these 3 variables are from 3 nested for loops.

more changes to tests to come


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@1409 dab4d1a6-ed17-0410-a064-d1ae371a2980