Purpose: bug fix
Description:
The fphdf5.o and similar files are not included in the make clean
target.  So, if the same directory is used again for a different
build, the left behind file could cause a problem (e.g., if
the second build is for a different binary format.)

Solution:
Added fphdf5.c to the TEST_SRC list which indirect form the clean
target list.

Platforms tested:
Did not run h5committest tests because they don't verify this
failure.  I tested the fix in modi4 pp to verify all files are cleaned.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@6718 dab4d1a6-ed17-0410-a064-d1ae371a2980