Purpose: Bug fix and feature
Description:
The patch to Makefile.in failed if an empty directory of fortran/examples
was left behind by a previous build.

Solution:
Attempted to put it in config/conclude.in just became a mess since
it is hard to predict when we will include examples in the subdirs
or not.
Borrow the structure of install-examples and do some handcode per
interface.  The Makefile.in of each interface (C, Fortran, C++)
controls everything.  Not very systematic but pretty clean for this
case.

Platforms tested:
h5committested.
Also tested in modi4 with various combinations of fortran, c++
enabled.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@6732 dab4d1a6-ed17-0410-a064-d1ae371a2980