Purpose: Bug fix
Description:
This would sometimes coredump in machines like PSC Alpha cluster
or linux icc if -g is not used.

Solution:
The real problem was because char string arrays were declared
without including space for the null terminator.  E.g.,
   char s[5][2]={ "s1", "s2", ...}
is not correct.  Increased the size and also need to change
the H5S... definitions too.

Platforms tested:
"h5committested"--the h5dumpgentest passed without problem though
it failed in Fortran but that is a current known problem.
Tested in verbena using icc and also in PSC's cluster.

Misc. update:
Can't find this error in the v1.5 know problem list.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@6946 dab4d1a6-ed17-0410-a064-d1ae371a2980