Purpose: "Bug fix"
Description:
This version is showing very slow HDF5 I/O write speeds (about 40%
of MPIO).  It was because the library is writing fill values by
default.  There is a no-fill option which set to FALSE.  Changed
it to TRUE since MPIO and POSIX are not doing fill values.
This is just a stop gap fix since the better fix would be to
change the library to not write fill-values in this case.

Platforms tested:
modi4 parallel and LANL Q paralllel.
No h5committest test because it is very trivia.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@6998 dab4d1a6-ed17-0410-a064-d1ae371a2980