Purpose: Bug fix.
Description:
There was a bug fix for AIX machine when Poe is used to
launch programs, that is in parallel mode. The fix is hostname
dependent.

Solution:
Removed the hostname dependence and let the fix apply to all
AIX machines.  But put in a safety net--if $HDF5_AIX_NOPATCH
is set to non-null, it will not use the fix but use the
original code.  (Don't expect this being used.)

Platforms tested:
Tested in Copper both with and without $HDF5_AIX_NOPATCH set.
Did not run h5committest since this really applies to AIX
systems only.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@7119 dab4d1a6-ed17-0410-a064-d1ae371a2980