Purpose:     Code cleanup, etc.
Description:
    Previously (in versions prior to 1.7), there were two locations to
modify when an error was added to the library.  Now, with the new error API,
there were four.

Solution:
    Created a single text file (src/H5err.txt) and a perl script (bin/make_err)
which uses the text file to automatically create header files that are included
in appropriate places in the library.
    This means that there is only one file (src/H5err.txt) which needs to be
modified when a new error code is added to the library.  The automatically
generated headers depend on this file and the makefiles will take care of
running the perl script to regenerate them when the text file changes, so no
user action is required when a new error is added.

Platforms tested:
    h5committested


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@7355 dab4d1a6-ed17-0410-a064-d1ae371a2980