Purpose: Improvement.
Description:
Complete change of the verbose control to use the routines provided by
the test/libh5test.a.
Also put in a temporary fix for the H5Eset_auto() and H5Eget_auto()
so that the Compat code are isolated in one place rather than all over
the source file.

Platforms tested:
Tested in Eirene (parallel).

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8096 dab4d1a6-ed17-0410-a064-d1ae371a2980