Purpose: bug fix.
Description:
On LANL QSC, test/dtypes "sw long double -> double" had failed when the
long double values were too small, smaller than the minimum double normalized
floating number.  The hw in QSC converts them mostly to 0 but occasionally,
it converts to the some non-zero values which are still <= the minimum
double normalized number.  But the conversion verification did not like
them and flagged them as errors.

Solution:
Added code to check if the src value is already smaller than the minimum
number the float class can hold.  When that happens and if both hw and sw
conversion results are <= the minimum number, then accepts them as okay
because if the src is smaller than the dst minimum number, an underflow
has occured.

Platforms tested:
Tested in LANL QSC and Theta.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8360 dab4d1a6-ed17-0410-a064-d1ae371a2980