Purpose:     Bug fix (sorta)
Description:
    The SGI machines have problems accurately (and consistently) converting
unsigned long values to float and double values, so put in a bit of a hack in
the datatype conversion test code to allow them to get "close enough".  This
hack is enabled at configure time by a flag which should only be set on machines
with this problem.

Platforms tested:
    FreeBSD 4.9 (sleipnir)
    h5committest


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8425 dab4d1a6-ed17-0410-a064-d1ae371a2980