Purpose:  Bug fix
Description:  For certain compiler(PGI we know so far), during conversion
from float or double to unsigned long long, it does round-up when the
fraction part is greater than 0.5, which shouldn't happen.

Solution:  check it during configuration and compensate this offset
during testing in dtypes.

Platforms tested:  verbena and fuss.  verbena is the only machine with
PGI compiler. Ran it on fuss to verify it with other compiler.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8491 dab4d1a6-ed17-0410-a064-d1ae371a2980