Purpose: Improvement
Description:
The MPI atomicity and file_sync tests may hang if a filesystem
is not able to support the operation.  This will block the
whole tests.  PHDF5 is not using either features.  So, removed
them from the default tests.


Platforms tested:
Only in eirene using pp.  Copper is still down.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8501 dab4d1a6-ed17-0410-a064-d1ae371a2980