Purpose: Bug Fix
Description:
Trying to create the root group or the working group ("/" or ".") fakes out
HDF5 so that it neither creates a group nor returns an error.

Solution:
H5G_namei now throws an error if it was supposed to insert but didn't.

Platforms tested:
sleipnir, Visual Studio 7 (very minor change)

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8897 dab4d1a6-ed17-0410-a064-d1ae371a2980