Purpose: Update.
Description:
Update all the hdf5 testing batch files.

Solution:
1. Updated dumptest.bat, difftest.bat, lstest.bat and repacktest.bat files
to match with correspondent testings in Unix.
2. Added one new batch file mask.bat to mask off time information displayed
in some h5ls tests.  The time displayed with h5ls uses a system all that
accountants for the local timezone of the cumputer that run that tests. To solve
this issue, the time information has to be masked off.  Otherwise, the expected
output may be different with the actual output.
3. Updated H5pubconf.h to enable szlib encoder.

Platforms tested:
Microsoft Visual C++ 6.0 on Windows XP and 2000.
(will test with .NET on Windows XP after this check-in.)

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@8925 dab4d1a6-ed17-0410-a064-d1ae371a2980