Purpose: Make collective chunk IO test more general
Description:
Previously collective chunk IO test is only fit for processor =4 with
the dimension size to be set small; sometimes people would like to test
with more than 4 processors(5,6 or more), the test therefore failed.
Solution:
To make the test case more general, dimensional size of the data is set to be large(right now 288 for each dimension), the disjoint hyperslab selection is re-calculated. Now the test cases should pass with 5,6 or 12 processors. Note, there is nothing wrong with the implementation of the library, it is the test case that causes the failure with the number of processor greater than 4.

Platforms tested:
Only at eirene, since only the test code is modified a little and it is very slow to test the parallel case.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9071 dab4d1a6-ed17-0410-a064-d1ae371a2980