Purpose: Feature
Description:
Added --private option which is for individuals making a private
release version.  It sets the SubRelease string to the date of
release.   This should be sufficient to distinguish releases
provided one does not make two private releases on the same
day.

Platforms tested:
No h5committest since it does not test this feature.
Hand tested it in Eirene.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9088 dab4d1a6-ed17-0410-a064-d1ae371a2980