Purpose: Bug fix.
Description:
Code would attempt to Calloc with zero count when a simple expression
that has no x term.  That resulted in NULL for some platform (like AIX).
That appeared as a failure treated as out of space.

Solution:
Checked if count is larger than 0 before making the calloc request.

Platforms tested:
Tested in copper (pp) where the failure appeared.  Also in eirene
as double check.  No h5committest as the change is trivial.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9297 dab4d1a6-ed17-0410-a064-d1ae371a2980