Purpose: Bug fix
Description:
Fix typo in #ifdef for unsigned long long to double conversion.
Changed WIN32 check to H5_ULLONG_TO_FP_CAST_WORKS
This check-in (and the one before it) also includes a large code-refactoring
that is still in progress.

Solution:
the ifdef for the plus case was checking unsigned long instead of unsigned
long long.  Moved the check to the right location.

Platforms tested:
sol+eirene

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9440 dab4d1a6-ed17-0410-a064-d1ae371a2980