Purpose: Code refactor Changed functionality of H5Pget_data_transform
Description:
Rewrote much of H5Ztrans.c to use macros, which cut down on the amount of code
significantly.

H5Pget_data_transform now allocates memory for the data transform string,
copies the string into this memory, and
returns this pointer to the user, instead of returning a pointer to the
internal transform string stored by hdf.  It is the user's responsibility to free
this memory when they are done with it.

Platforms tested:
sol + eirene + copper

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9441 dab4d1a6-ed17-0410-a064-d1ae371a2980