Purpose: 1.Support Fortran DLL for Decfortran 6.6c. 2. misc. small bug fixes.
Description:
1. No HDF5 fortran DLL is provided in the recent HDF5 release because of
     1) the difficulties to build fortran DLL based on C DLL.
     2) the difficulties to separate fortran DLL from C DLL in hdf5 test suite.
     3) DEC Fortran 6.0 was old and we don't have new DEC fortran 6.6 compiler.


Solution:
0. Updating DEC Fortran 6.6A compiler to 6.6c since there was a bug in 6.6A to fail export variable within module procedure.
1. Using a C stub dll to separate fortran and C codes, previously they were mixed.
2. Using a C stub dll for test suite.
3. Carefully passing the token between export and import DLLs.
4. Now Fortran tests still failed with 1.7 branch, however,
the failure is due to bugs inside HDF5 library because both static and DLL experienced the same failure.
The failure occurred for routine h5premove_f. According to the failure message, it may be due to the invalid array bound.
Anyway, Fortran DLL can be treated as the same as the static library for 1.7 branch with DEC visual 6.6c.


Platforms tested:
windows xp with DEC fortran 6.6c and VS 6.0.

Misc. update:
update some batch files and project file related to this change.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9473 dab4d1a6-ed17-0410-a064-d1ae371a2980