Purpose:  Bug fix
Description:  the last 2 bytes of mantissa can be lost when converting from unsigned long long
to long double.  In last check-in, a fix was made to ignore that precision loss.  But sometimes,
the last 2 bytes are rounded up to the 3rd last byte.

Solution:  Ignore the 3rd last byte, too, when comparing values.

Platforms tested: sleipnir - only this system is concerned.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9814 dab4d1a6-ed17-0410-a064-d1ae371a2980