Purpose:  Bug fix for test program
Description:
For Intel machines, the size of "long double" is 12 byte, precision
is 80 bits, mantissa size is 64 bits, and no normalization.  So the
most significant bit of mantissa is always 1 unless the floating number
has special value.  This step tries to compensate this case by turning
on the most significant bit of mantissa if the mantissa bits aren't
all 0s.


Solution:
Tries to compensate this case by turning on the most significant bit of
mantissa if the mantissa bits aren't all 0s.

Platforms tested: eirene and fuss(production enabled).  Small change for
production only.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9898 dab4d1a6-ed17-0410-a064-d1ae371a2980