Purpose: Bug fix
Description:
h5diff sometimes will 'exit(status)' but if it is in parallel mode,
it needs to close up the workers and the MPI environment before exit(..).

Solution:
Created h5diff_exit() that will just exit in serial mode but does all
the shutdown properly if in parallel mode.

Platforms tested:
tested in heping, serial and parallel.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@9949 dab4d1a6-ed17-0410-a064-d1ae371a2980