Purpose: Added Packet Table to high-level APIs
Description:
The Packet Table is an API that allows the user to append records
("packets") to a table, and read the back again.  It supports
fixed-length records with a defined datatype and variable-length
records.  It also supports a "current record" index to track the
user's position in the table.

Solution:
The Packet Table code lives in hl/src, and its tests in hl/test.
Some code is shared between the H5TB table and the H5PT Packet Table
in the form of functions in H5HL_private.c.
Some documentation exists for a previous version of the API.  Updated
documentation and C++ wrapper API coming soon.

Platforms tested:
sleipnir, eirene, copper, modi4


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10234 dab4d1a6-ed17-0410-a064-d1ae371a2980