Purpose: Bug fix.
Description:
Data file name is hardcoded to be in current directory which does not
necessarily support MPIO. Changed it to /tmp/sds.h5 for slightly larger
chance of success but the eventual solution is to set it according to
environment variables.  Patch it this way for now.

Platforms tested:
mir.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10593 dab4d1a6-ed17-0410-a064-d1ae371a2980