Purpose: Feature - libtool shared library versioning
Description:
Libtool provides a mechanism for different versions of the same shared
library to be distinguished.  Now this is applied to HDF5 when it is
built as a shared library.

Solution:
The version number is stored in config/lt_vers.am, and included in
src/Makefile.am.
This number will be automatically updated by bin/h5vers; developers
only need to update it when they change the API.

*** IMPORTANT ***
Any time the API changes, the version number in config/lt_vers.am must
be updated!



Platforms tested:
mir, eirene, verbena, modi4

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10627 dab4d1a6-ed17-0410-a064-d1ae371a2980