Purpose: Update windows fortran support
Description:
Currently an implementation of automatic generated fortran data type from C has been done for
Linux/Unix platforms. For windows different implementation needs to be developed to make it
consistent with the implementation on Linux/Unix. For the purpose of 1.8 release, we have to add a
new header file. The checking program of the header file on windows will be generated on windows separately.
Solution:

Platforms tested:
VS 6.0 on windows xp.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10663 dab4d1a6-ed17-0410-a064-d1ae371a2980