Purpose: Bug fix
Description:
Description: See details from Bug #213.  Family member file size wasn't saved
anywhere in file.  When family file is opened, the first member size determine
the member size.

Solution:  This is the second step of fixing this bug.  Multi driver is checked against the driver
name saved in superblock.  Wrong driver will result in a failure with an error message indicating
multi driver should be used.  This change includes split driver because it's a special case for multi
driver.

In the first step of checkin.  Family member size and name template(unused at this stage) are saved
in file superblock.  When file is reopened,the size passed in thrin superblock.  A different size
will trigger a failure with an error message indicating the right size.  Wrong driver to open family
file will cause a failure, too.

The third step will be changing h5repart to let it change family member size in the superblock if
user uses it to change member file size.

Platforms tested: h5committest and fuss.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10778 dab4d1a6-ed17-0410-a064-d1ae371a2980