Purpose: Configuration feature
Description:
Formerly, bin/reconfigure needed to invoke Automake with the --foreign
flag or it would fail.  It turns out that there is a clever way to
specify this flag inside configure.in so that it doesn't need to
be explicitly invoked.
This should make it easier for people not using machines connected
to AFS (e.g., the netCDF folks) to invoke Automake on HDF5.


Platforms tested:
mir and heping


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10802 dab4d1a6-ed17-0410-a064-d1ae371a2980