Purpose: Solve the Difference of Compiler and Library
Description: In this data conversion test, from floating number to floating number,
when the source is bigger than the destination, overflow may happen.  The library
assign INFINITY while compiler's behavior is undefined.  A compiler like SGI assigns
the maximal value for the destination.

Solution: When this situation occurs, simply check if the destination is INFINITY
instead of comparing values.

Platforms tested: modi4 and h5committest.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@10962 dab4d1a6-ed17-0410-a064-d1ae371a2980