Purpose: tidy up.
Description:
MPE option created *.clog file whenever it is executed in MPI.
The cleanup of *.clog files were done in individual Makefile.in.
Often, it is forgotten.

Solution:
Moved the cleaning of *.clog files to CHECK_CLEANFILES in commence.am
so that it is applied whenever check-clean is called.

Platforms tested:
heping pp using MPE.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11323 dab4d1a6-ed17-0410-a064-d1ae371a2980