Purpose: Feature - md5 checksum
Description:
Added md5 checksumming to snapshot releases.  When bin/release is run,
it will by default create a .tar archive and a .md5 checksum of that
archive in the destination directory.

Solution:
md5 is now a method in bin/release, along with the various ways to compress
hdf5.  The checksum is always of the .tar archive, even when other forms
of compression are used (this is so that a given release always has only
one checksum, for simplicity).

Platforms tested:
mir (change to bin/release only)

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11362 dab4d1a6-ed17-0410-a064-d1ae371a2980