Purpose: Improvement.
Description:
The test may hang if there are system failures that some processors
are not working.

Solution:
Added the ALARM calls to limit all tests be done with the default
alarm time.  So, even if a process is hanging, the ALARM signal
would terminate the process.

Platforms tested:
tg-login2 of NCSA.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11458 dab4d1a6-ed17-0410-a064-d1ae371a2980