Purpose: bug fix.
Description:
When a parallel test script test fails, make would continue because the
way it was setup inside a for loop.  Fixed it by issuing an exit 1 inside
the loop.

There was also a typo error in the newer command comparision that it
must be $${chkname} in order to be valid.  Also, the test script itself
was not checked in the newer lists.  All fixed.

Platforms tested:
h5committested and also hand tested in heping pp mode.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11478 dab4d1a6-ed17-0410-a064-d1ae371a2980