Purpose: File cleanup.
Description:
The file contained \r.  Looked like it was created in Windows
and commited as is.

Solution:
Removed all \r and diff with original with -w to verify no code
changes.

Platforms tested:
This file was not used in the Makefile at all.  Compiled it by
hand using h5cc.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11523 dab4d1a6-ed17-0410-a064-d1ae371a2980