Purpose: Cleanup
Description:
Changed comments in C++ packet table code to refer to entries as
"packets" instead of "records."

Solution:
The APIs were changed to correspond to the C API, but the documentation
wasn't.  This is fixed.

Platforms tested:
mir, shanti, sleipnir

Misc. update:
This corresponds to documentation update earlier this week.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11751 dab4d1a6-ed17-0410-a064-d1ae371a2980