Purpose: Cleanup
Description:
v1.7:
Removed some debug files left by previous change.
Uncommented the removal of the temporary working directory.

v1.6:
Moved the code of removal of the temporary working directory
to a later location in order to stay in sync with v1.7.
(The later location is needed for v1.7 code but makes no different
to v1.6)

Platforms tested:
Tested by hand.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@11935 dab4d1a6-ed17-0410-a064-d1ae371a2980