Purpose: Configuration change
Description:
Stream-VFD is now disabled when parallel is enabled.

Solution:
Stream-VFD was causing trouble on some platforms when both it and
parallel HDF5 were being used.  This is a pretty odd use case, so
we now disable stream vfd when parallel is being used.
This required moving the test for the stream vfd after the test
for parallel HDF5 in configure.in.

Platforms tested:
mir, modi4 (configure change only)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@12230 dab4d1a6-ed17-0410-a064-d1ae371a2980