Purpose: Follow-up what's committed yesterday.
Description: Yesterday, a definition of the macro YY_BUF_SIZE of 256KB
was committed in, but copper complained it can't be redefined.

Solution: Added #ifdef and #undef to make sure the macro is undefined
first.  Also integrated a complicated compound type test.  It reads
input from a text file.  The datatype is unusually large, of 101 member
fields.

Platforms tested: h5committest and fuss.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@12351 dab4d1a6-ed17-0410-a064-d1ae371a2980