Purpose: Bug fix.
Description:
Sometimes the parallel prefix is given in the form of nfs:/mnt/pfs which
if given to a non-MPIO VFD like the default H5Fcreate, it would fail.
Called h5_rmprefix which would return the non-prefix component
part of the file name which would be okay for the default H5Fcreate and
such.

Platforms tested:
Tested in heping parallel.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@12396 dab4d1a6-ed17-0410-a064-d1ae371a2980