Purpose: Bug fix (related to 544)
Description:
h5_get_file_size() was coded to return 0 if failed but
file size can be 0.  Changed the failure return value to -1
which is allowed by the returned type of off_t which is a signed
type.
Also changed the checking code of the stat call to just == 0
since that is how it is defined.

Test:
Could test it in heping only.  Both Sol and Copper failed to
compiled due to error in the Direct IO VFD code.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@12751 dab4d1a6-ed17-0410-a064-d1ae371a2980